+++ packages/core/src/component/index.ts
@@ -1,0 +1,3 @@
+export * from '../services/component.service';
+export * from '../component/component';
+export * from '../component/riba-component';