+++ packages/core/src/components/index.ts
@@ -1,3 +1,1 @@
-export * from '../services/component.service';
-export * from './abstract-riba-component';
-export * from './riba-component';
+export { VideoComponent } from './video/video.component';